refactor: move logger to log mod
